+++ frontend/src/Editor/Components/BoundedBox/Box.jsx
@@ -1,0 +1,17 @@
+import React from 'react';
+export const Box = ({ children, geometry, style }) => {
+  return (
+    <div
+      style={{
+        ...style,
+        position: 'absolute',
+        left: `${geometry.x}%`,
+        top: `${geometry.y}%`,
+        height: `${geometry.height}%`,
+        width: `${geometry.width}%`,
+      }}
+    >
+      {children}
+    </div>
+  );
+};
