+++ frontend/src/Editor/Components/BoundedBox/RenderSelector.jsx
@@ -1,0 +1,17 @@
+import React from 'react';
+import { Box } from './Box';
+// eslint-disable-next-line no-unused-vars
+export const RenderSelector = ({ annotation, active }) => {
+  let { geometry } = annotation;
+  if (!geometry) return null;
+
+  return (
+    <Box
+      geometry={geometry}
+      style={{
+        background: 'rgba(255, 255, 255, 0.5)',
+        border: 'solid 1px red',
+      }}
+    ></Box>
+  );
+};
